Fix database generation from EF (#264)
* fix bug with generation db by ef

* update fix

* change CourseId type from "long?" on "long"

* fix bug after update 'dev'

+++ CharlieBackend.Data/Repositories/Impl/DashboardRepository.cs
@@ -161,7 +161,7 @@ namespace CharlieBackend.Data.Repositories.Impl
                     .Select(x => new
                     {
                         CourseId = x.Lesson.StudentGroup.CourseId,
-                        StudentGroupId = (long)x.Lesson.StudentGroupId,
+                        StudentGroupId = x.Lesson.StudentGroupId,
                         StudentLessonMark = x.StudentMark,
                         StudentId = x.StudentId
                     })
